phase-functions: Rework INSTALL_MASK to apply as a merge filter
Rework the INSTALL_MASK handler to filter installed files while merging
instead of removing them from the installation image.

The INSTALL_MASK handling is now moved to Python side of code, with bash
being only used to evaluate the value of INSTALL_MASK (with respect to
bashrc).

The evaluated value of INSTALL_MASK is used to test for collisions,
and afterwards to skip masked files from being installed. It is also
used in uninstall code to properly remove newly-masked files that were
installed previously.