AbstractEbuildProcess: use async_unlock (bug 614108)
Override the _async_wait method to asynchronously unlock
self._build_dir when necessary, override the _wait method in order
to function as a failsafe if _async_wait has not been called for
some reason, and fix the SubProcess superclass to call _async_wait
when appropriate.

Execution of the _wait method's failsafe code will automatically
become a fatal error at the same time as event loop recursion is
disabled.

Bug: https://bugs.gentoo.org/614108