introduce the 'stacked-prefix' FEATURE
When we merge into another EPREFIX, but not into some ROOT,
and CHOST is equal to CBUILD, build tools found in EPREFIX
perfectly work for the current build environment.
In a "stacked prefix" we explicitly utilize this situation.

This is useful during prefix bootstrap (#655414, #655326), but also to
build packages for targets unable to support the full portage toolchain
(native Windows, MinGW), but otherwise do not require a full cross
compilation setup.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>