portage.data: add build user (bug 600804)
Add attributes and override variables for a separate build
user, with defaults corresponding to the status quo. This
reuses the existing PORTAGE_BUILD_USER and PORTAGE_BUILD_GROUP
variables that were already exported to the ebuild environment.

When PORTAGE_BUILD_USER and PORTAGE_BUILD_GROUP are unset,
the build user can still be overridden by the existing
PORTAGE_USERNAME and PORTAGE_GRPNAME variables.

This sets the stage for migration of internal code to use
a separate build user, as required to solve bug 600804 in a
manner that complies with PMS.

Bug: https://bugs.gentoo.org/600804