Revert "install-qa-checks.d: Add a check for Gentoo path policies (FHS-y)"
This reverts commit d5f97eaa464736a454c8ad410f4acd4fccdf2324.
Until this QA check has adjustable whitelist support, we can consider
it an unstable work in progress. Therefore, I'd like for the QA team
to move it gentoo/metadata/install-qa-check.d/08gentoo-paths until
it has matured. It's safe to commit it to the gentoo repository now,
and it will become active when portage's internal copy is removed.

Bug: https://bugs.gentoo.org/667604
Closes: https://bugs.gentoo.org/667378
Closes: https://github.com/gentoo/portage/pull/373
Signed-off-by: Zac Medico <zmedico@gentoo.org>