man: Remove obsolete cross-compilation section
It makes heavy use of ambiguous terminology from EAPI 5-hdepend. Part
of it talks about the targetroot USE flag, which we do not plan to
formally adopt into PMS. Overall, it is more likely to confuse than
help people.

I thought about rewriting it but it is difficult to summarise this
complex subject in a useful way. The wiki is probably a better place
to write about it.

Closes: https://github.com/gentoo/portage/pull/413
Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Signed-off-by: Zac Medico <zmedico@gentoo.org>