autounmask: prefer || choices with existing packages (bug 327177)
When autounmask is enabled, make depgraph retry _select_atoms calls when
necessary to prefer || choices with existing packages. For example,
if package C is masked and package B does not exist, then autounmask
should choose C when given the choice || ( B C ), as shown in the
included unit tests. The unit tests also show that autounmask still
prefers choices containing packages that are not masked (if available).

Bug: https://bugs.gentoo.org/327177
Signed-off-by: Zac Medico <zmedico@gentoo.org>