emerge: Don't treat empty EPREFIX or PORTAGE_CONFIGROOT as unset
If a prefix user wanted to build within a ROOT but without a prefix,
they previously had to set EPREFIX=/ rather than EPREFIX= as the
latter was simply treated as unset.

Also applies to ROOT and SYSROOT but probably makes no difference to
these as they are blank by default. This should be safe to do as all
these variables get normalised anyway.

Bug: https://bugs.gentoo.org/642604
Closes: https://github.com/gentoo/portage/pull/434
Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Signed-off-by: Zac Medico <zmedico@gentoo.org>