Add FEATURES=pkgdir-index-trusted
Trust that the PKGDIR index file is valid, meaning that no packages
have been manually added or removed since the last call to emaint --fix
binhost. This feature eliminates overhead involved with detection of
packages that have been manually added or removed, which significantly
improves performance in some cases, such as when PKGDIR resides on
a high-latency network file system.

Bug: https://bugs.gentoo.org/688902
Signed-off-by: Zac Medico <zmedico@gentoo.org>