man/ebuild.5: Improvements to the ebuild(5) man page
The following patch contains several formatting improvements,
as well as extended information about KEYWORDS that is easier
to understand. The original description started describing Gentoo
policy on KEYWORDS without actually explaining what they are and how
they work.

I have also added missing documentation for the ARCH variable as this
was sort of shoe-horned in to the KEYWORDS documentation, and tried
to make it clearer what is a technical requirement of Portage the
software and what are Gentoo-specific arch team requirements.

Bug: https://bugs.gentoo.org/695870
Signed-off-by: Zac Medico <zmedico@gentoo.org>