emerge: Log completion of package installs
The log currently shows Emerging/Installing progress messages, but
it doesn't indicate when a package finishes.  Since there can be dozens
of installs in flight at once, this makes it nearly impossible to
determine how long an individual package took.  This adds a similar
"Completed" line when the package merge completes.

Closes: https://github.com/gentoo/portage/pull/987
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>
Signed-off-by: Matt Turner <mattst88@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>